demo-app/requirements.txt
flask
opentelemetry-sdk
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp
opentelemetry-api
# opentelemetry-instrumentation
